Adding Type definitions for the Kit.json
I was receiving cache issues when running development for a second time, not a big deal locally but this was breaking the CI when I added the project to first Netlify then Gatsby Cloud. The error message suggested adding the Types for the data explicitly and lo and behold following that advice fixed the problem!